+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 
 ## Unreleased
 
+### Added
+
+- Added possibility to manipulate serverside ConVars with Dropdowns
+  - Just add .serverConvar with the conVarName to the given data similar to .convar
+
 ## [v0.11.3b](https://github.com/TTT-2/TTT2/tree/v0.11.3b) (2021-11-18)
 
 ### Fixed